% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcrq.R
\name{dcrq}
\alias{dcrq}
\title{Fit the doubly interval-censored AFT model with quantile regression}
\usage{
dcrq(
  L,
  R,
  T,
  delta,
  x,
  tau,
  estimation = NULL,
  wttype = "param",
  hlimit = 0.5,
  id = NULL,
  k = 1,
  maxit = 100,
  tol = 0.001
)
}
\arguments{
\item{L}{left-censoring time, having 0 if left-censored.}

\item{R}{right-censoring time, having \code{Inf} if right-censored.}

\item{T}{exactly observed time.}

\item{delta}{censoring indicator, 1: observed; 0: interval-censored.}

\item{x}{X matrix of baseline covariates.}

\item{tau}{quantile level.}

\item{estimation}{estimating method of partly interval censored, if estimation="dr", doubly robust estimator is estimated.}

\item{wttype}{weight estimating method, default is "param" and Beran's nonparametric KM estimating method as "nonparam".}

\item{hlimit}{bandwidth value, default is 0.5.}

\item{id}{cluster id. If the data does not have clustered structure, set \code{id=NULL}.}

\item{k}{index of cluster weight.}

\item{maxit}{maximum number of iteration for the log-rank estimator, default is 100.}

\item{tol}{tolerance of iteration for the log-rank estimator, default is 1e-3.}
}
\value{
\code{dcrq} returns a data frame containing at least the following components:
\itemize{
\item \code{tau}: quantile level.
\item \code{coefficients}: regression estimator.
\item \code{se}: standard error estimates for \code{est}.
\item \code{pvalue}: p-value.
\item \code{lower bd}: lower bound of coefficients under 95\% confidence level.
\item \code{upper bd}: upper bound of coefficients under 95\% confidence level.
}
}
\description{
Fit inverse weighted quantile regression with doubly interval-censored data
}
\details{
see Kim et al., (2022+) for detailed method explanation.
}
\examples{
\dontrun{
# Simulations
n=200; x1=runif(n,-1.2,1.7); x2=rbinom(n,1,0.6)
T = 1.7+x1+x2+rnorm(n)*(1-0.1*x2)
L=runif(n,-2.8,1.9); R=L+runif(n,4.2,8.1)
Y=pmin(R,pmax(T,L))
delta=case_when(
 T<L ~ 3,
 T>R ~ 2,
 TRUE ~ 1 #observed
)
L=L; R=R; T=T; delta=delta; x=cbind(x1,x2)
dcrq(L,R,T,delta,x,tau)
dcrq(L,R,T,delta,x,tau,estimation = "dr")
dcrq(L,R,T,delta,x,tau,wttype = "nonparam")
dcrq(L,R,T,delta,x,tau,wttype = "nonparam",hlimit = 0.9)
}
}
\references{
Beran, R. (1981). Nonparametric Regression with Randomly Censored Survival Data. Technical Report, Univ.California, Berkeley.

Chiou, S. H., Kang, S. and Yan, J. (2015). Rank-based estimating equations with general weight for accelerated failure time models: an induced smoothing approach. Statistics in Medicine 34(9): 1495–-1510.

Kim, Y., Choi, T., Park, S., Choi, S. and Bandyopadhyay, D. (2022+). Inverse weighted quantile regression with partially interval-censored data.

Pan, C. (2021). PICBayes: Bayesian Models for Partly Interval-Censored Data. R package. https://CRAN.R-project.org/package=PICBayes.
}
