% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picrq.R
\name{picrq}
\alias{picrq}
\title{Fit the partly interval-censored AFT model with quantile regressions}
\usage{
picrq(
  L,
  R,
  delta,
  x,
  tau,
  estimation = NULL,
  wttype = "param",
  h = 0.5,
  id = NULL,
  k = 1,
  maxit = 100,
  tol = 0.001
)
}
\arguments{
\item{L}{left-censoring time, having 0 if left-censored.}

\item{R}{right-censoring time, having \code{Inf} if right-censored.}

\item{delta}{censoring indicator, 1: observed; 0: interval-censored.}

\item{x}{baseline covariates.}

\item{tau}{quantile level.}

\item{estimation}{estimating method of partly interval censored, if estimation="dr", doubly robust estimator is estimated.}

\item{wttype}{weight estimating method, default is "param".}

\item{h}{bandwidth value, default is 0.5.}

\item{id}{cluster id. If the data does not have clustered structure, set \code{id=NULL}.}

\item{k}{index of cluster weight.}

\item{maxit}{maximum number of iteration for the log-rank estimator, default is 100.}

\item{tol}{tolerance of iteration for the log-rank estimator, default is 1e-3.}

\item{T}{exactly observed time.}
}
\value{
\code{picrq} returns a data frame containing at least the following components:
\itemize{
\item \code{coefficients}: regression estimator.
\item \code{se}: standard error estimates for \code{est}.
\item \code{pvalue}: p-value.
\item \code{lower bd}: lower bound of coefficients under 95\% confidence level.
\item \code{upper bd}: upper bound of coefficients under 95\% confidence level.
}
}
\description{
Fit inverse weighted quantile regression with partially interval-censored data
}
\details{
see Kim et al., (2022+) for detailed method explanation.
}
\examples{
\dontrun{
# Simulations
# library(tidyverse)
library(ipcwQR)
set.seed(111)
x1 = rnorm(n)
x2 = rbinom(n,1,0.5)
x = cbind(x1,x2)
T = 2 + x1 + x2 + rnorm(n)
U = (1 - 0.25*x1)*runif(n, -6, 3.85)
V = U + (1 - 0.5*x2)*runif(n, 6, 19.6)
U = exp(dplyr::case_when(TRUE ~ T, T>V ~ V, T<U ~ -Inf))
V = exp(dplyr::case_when(TRUE ~ T, T>V ~ Inf, T<U ~ U))
delta = ifelse(U==V, 1, 0); tau=0.3
picrq(V,U,delta,x,tau)
picrq(V,U,delta,x,tau,estimation = "dr")


# Data example
library(PICBayes)
library(ipcwQR)
data(mCRC)
d=data.frame(mCRC)
y=d$y; U = ifelse(y==0,R,L); V = ifelse(y==2,L,R)
delta = case_when(d$IC == 0 ~ 1, d$IC == 1 ~ 0)
x1= case_when(d$TRT_C == 0 ~ 0, d$TRT_C == 1 ~ 1)
x2= case_when(d$KRAS_C == 0 ~ 1, d$KRAS_C == 1 ~ 0)
x = cbind(x1,x2);  
id = rep(c(table(d$SITE)),c(table(d$SITE))); tau=0.3
picrq(V,U,delta,x,tau)
picrq(V,U,delta,x,tau,wttype = "nonparam",h=0.9)
picrq(V,U,delta,x,tau,wttype = "nonparam",id=id,h=0.9)


d = with(dt0,
         data.frame(U = ifelse(is.na(L), 0, L),
                    V = ifelse(is.na(R), Inf, R),
                    Delta = 1-IC,
                    x1 = TRT_C,
                    x2 = KRAS_C,
                    id = SITE))
U = d$U; V = d$V; X = cbind(d$x1,d$x2); Delta = d$Delta; id = d$id
aft_rank(U = U, V = V, X = X, Delta = Delta, id = id, 
         alpha = 1, type = "gehan", nboot = 10)
aft_rank(U = U, V = V, X = X, Delta = Delta, id = id, 
         alpha = 1, type = "logrank", nboot = 10)
}
}
\references{
Chiou, S. H., Kang, S. and Yan, J. (2015). Rank-based estimating equations with general weight for accelerated failure time models: an induced smoothing approach. Statistics in Medicine 34(9): 1495–-1510.

Kim, Y., Choi, T., Pak, S., Choi, S. and Bandyopadhyay, D. (2022+). Inverse weighted quantile regression with partially interval-censored data.

Pan, C. (2021). PICBayes: Bayesian Models for Partly Interval-Censored Data. R package. https://CRAN.R-project.org/package=PICBayes.
}
