% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcrq.R
\name{dcrq}
\alias{dcrq}
\title{Fit the doubly interval-censored AFT model with quantile regression}
\usage{
dcrq(
  L,
  R,
  T,
  delta,
  x,
  tau,
  estimation = NULL,
  wttype = "param",
  hlimit = 0.5,
  id = NULL,
  k = 1,
  maxit = 100,
  tol = 0.001
)
}
\arguments{
\item{L}{left-censoring time, having 0 if left-censored.}

\item{R}{right-censoring time, having \code{Inf} if right-censored.}

\item{T}{exactly observed time.}

\item{delta}{censoring indicator, 1: observed; 0: interval-censored.}

\item{x}{baseline covariates.}

\item{tau}{quantile level.}

\item{estimation}{estimating method of partly interval censored, if estimation="dr", doubly robust estimator is estimated.}

\item{wttype}{weight estimating method, default is "param".}

\item{id}{cluster id. If the data does not have clustered structure, set \code{id=NULL}.}

\item{k}{index of cluster weight.}

\item{maxit}{maximum number of iteration for the log-rank estimator, default is 100.}

\item{tol}{tolerance of iteration for the log-rank estimator, default is 1e-3.}

\item{h}{bandwidth value, default is 0.5.}
}
\value{
\code{dcrq} returns a data frame containing at least the following components:
\itemize{
\item \code{coefficients}: regression estimator.
\item \code{se}: standard error estimates for \code{est}.
\item \code{pvalue}: p-value.
\item \code{lower bd}: lower bound of coefficients under 95\% confidence level.
\item \code{upper bd}: upper bound of coefficients under 95\% confidence level.
}
}
\description{
Fit inverse weighted quantile regression with doubly interval-censored data
}
\details{
see Kim et al., (2022+) for detailed method explanation.
}
\examples{
\dontrun{
# Simulations
rm(list=ls())
library(tidyverse)
set.seed(111)
cluster=60
v=rlnorm(cluster,mean=0,sd=0.5)
k = sample(0:9*10,cluster,replace = TRUE)
m = ifelse(quantile(v, k/100) <= v & quantile(v, (k+10)/100) <= v,
           k/10 + 2, 2)
id = as.numeric(unlist(apply(cbind(1:cluster,m),1,function(a) rep(a[1],a[2]))))
n=sum(m)
x1 = rnorm(n); x2 = rbinom(n,1,0.5); x = cbind(x1,x2)
T = exp(2 + x1 + x2 + rnorm(n))
L=runif(n,-2.8,1.7); R=L+runif(n,4.2,10); 
L=exp(L); R=exp(R)
Y=pmin(R,pmax(T,L))
delta=case_when(T<L ~ 3, T>R ~ 2, TRUE ~ 1) #observed
tau=0.3
dcrq(L,R,T,delta,x,tau)
}
}
\references{
Chiou, S. H., Kang, S. and Yan, J. (2015). Rank-based estimating equations with general weight for accelerated failure time models: an induced smoothing approach. Statistics in Medicine 34(9): 1495–-1510.

Kim, Y., Choi, T., Park, S., Choi, S. and Bandyopadhyay, D. (2022+). Inverse weighted quantile regression with partially interval-censored data.

Pan, C. (2021). PICBayes: Bayesian Models for Partly Interval-Censored Data. R package. https://CRAN.R-project.org/package=PICBayes.
}
